refactoring file structure, CD deployment not updated yet
